Addresses comments made by Nick in the PR. Added some more comments. Made fromList take a combining function to address what to do with duplicate keys. Split off Control.Iterate.SetAlgebraInternal.hs so that Control.Iterate.SetAlgebra can export only those things needed to create and run Set expressions.
