+++ semantics/executable-spec/small-steps.cabal
@@ -28,6 +28,7 @@ library
                      , Data.AbstractSize
                      , Control.Iterate.Collect
                      , Control.Iterate.SetAlgebra
+                     , Control.Iterate.SetAlgebraInternal
   -- other-modules:
   -- other-extensions:
   build-depends:       base >=4.11 && <5
@@ -37,7 +38,6 @@ library
                      , free
                      , mtl
                      , transformers >= 0.5
-                     , QuickCheck
                      , ansi-wl-pprint
                      -- IOHK deps
                      , cardano-crypto-class
